feat: 구간 관리

+++ src/test/java/subway/StationAcceptanceTest.java
@@ -2,15 +2,12 @@ package subway;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ActiveProfiles;
 import subway.fixture.StationTestFixture;
-import subway.setup.DataBaseCleanUp;
 
 import java.util.List;
 
@@ -81,7 +78,7 @@ public class StationAcceptanceTest extends BaseTest{
     @DisplayName("지하철역을 삭제한다.")
     @Test
     void deleteStations(){
-        long id = StationTestFixture.createStationFromName(지하철_이름_강남역).jsonPath().getInt("id");
+        long id = StationTestFixture.createStationFromName(지하철_이름_강남역).jsonPath().getLong("id");
         StationTestFixture.deleteById(id);
 
         List<Long> ids = StationTestFixture.allStations().jsonPath().getList("id", Long.class);
